requirements.txt
Flask==2.3.3
gunicorn
psycopg2-binary==2.9.11

